archive/ai_analysis.py
# archive/ai_analysis.py
# (Originally src/modules/ai_analysis.py)
import torch
import torch.nn as nn

# Note: This appears to be a standalone example using PyTorch for a GAN.
# Requires PyTorch: See https://pytorch.org/get-started/locally/
# Unrelated to the main simulation core or the TensorFlow-based AI tool.

class MaliciousGAN(nn.Module):
    """
    Demonstrates a simple GAN-like architecture for generating binary data
    that could be used as an obfuscated payload in authorized red-team exercises.
    """

    def __init__(self, input_dim=100, output_dim=512):
        super().__init__()
        self.input_dim = input_dim
        self.generator = nn.Sequential(
            nn.Linear(input_dim, 256),
            nn.LeakyReLU(0.2, inplace=True),
            nn.Linear(256, output_dim),
            nn.Tanh() # Output range [-1, 1]
        )
    
    def generate_payload_data(self, num_samples=1, noise=None):
        """
        Generates mock binary data from random noise.
        Output is raw bytes, interpretation depends on context.
        This is for demonstration and testing in isolated labs only.
        """
        if noise is None:
            # Generate noise for the specified number of samples
            noise = torch.randn(num_samples, self.input_dim)
        elif noise.shape[0] != num_samples or noise.shape[1] != self.input_dim:
             raise ValueError(f"Provided noise shape {noise.shape} incompatible with num_samples={num_samples} and input_dim={self.input_dim}")

        self.generator.eval() # Set model to evaluation mode
        with torch.no_grad():
            # Generate data in the [-1, 1] range
            generated_data_tensor = self.generator(noise)
            # Scale to [0, 255] byte range
            generated_data_scaled = ((generated_data_tensor + 1) / 2.0 * 255.0).round().byte()
            
        # Convert to a list of byte strings if num_samples > 1
        if num_samples == 1:
            return generated_data_scaled.cpu().numpy().tobytes()
        else:
            return [sample.cpu().numpy().tobytes() for sample in generated_data_scaled]

# Example Usage (if run directly)
if __name__ == '__main__':
    gan = MaliciousGAN()
    payload = gan.generate_payload_data()
    print(f"Generated {len(payload)} bytes of data.")
    # print(payload.hex())

    # Generate multiple payloads
    payloads = gan.generate_payload_data(num_samples=3)
    for i, p in enumerate(payloads):
         print(f"Payload {i+1}: {len(p)} bytes") 