requirements.txt
# Core Dependencies
pyyaml>=6.0.1
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0
scapy>=2.5.0
paramiko>=3.3.0
dnspython>=2.4.0
netifaces>=0.11.0
pyOpenSSL>=23.2.0
pycryptodome>=3.19.0

# Network and Protocol
aiohttp>=3.9.0
websockets>=12.0
socks>=0.3.0
python-nmap>=0.7.1
pyshark>=0.6.0

# Security and Encryption
pycryptodomex>=3.19.0
pyjwt>=2.8.0
certifi>=2023.11.17
urllib3>=2.1.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# Utilities
tqdm>=4.66.1
colorama>=0.4.6
rich>=13.7.0
click>=8.1.7

# Add platform-specific dependencies
pywin32>=306; sys_platform == 'win32' # For Windows API access (Parent PID Spoofing, etc.)

# Removed Flask>=2.2.0 (Potential future UI/API, not currently used)
# Removed pycuda (Used by archived archive/gpu_payload.py)